+++ cpp/opendnp3tests/src/MockPhysicalLayerAsync.cpp
@@ -78,7 +78,7 @@ void MockPhysicalLayerAsync::SignalOpenFailure()
 
 void MockPhysicalLayerAsync::SignalSendSuccess()
 {
-	size_t num = mNumToWrite; mNumToWrite = 0;
+	uint32_t num = mNumToWrite; mNumToWrite = 0;
 	std::error_code ec;
 	this->OnWriteCallback(ec, num);
 }
